+++ HomegearLib.NET/HelperFunctions.cs
@@ -4,16 +4,16 @@ namespace HomegearLib
 {
     public static class HomegearHelpers
     {
-        public static DateTime UnixTimeStampToDateTime(int unixTimeStamp)
+        public static DateTime UnixTimeStampToDateTime(long unixTimeStamp)
         {
             System.DateTime epoch = new DateTime(1970, 1, 1, 0, 0, 0, 0, System.DateTimeKind.Utc);
             return epoch.AddSeconds(unixTimeStamp).ToLocalTime();
         }
 
-        public static int DateTimeToUnixTimeStamp(DateTime time)
+        public static long DateTimeToUnixTimeStamp(DateTime time)
         {
             System.DateTime epoch = new DateTime(1970, 1, 1, 0, 0, 0, 0, System.DateTimeKind.Utc);
-            return (int)time.ToUniversalTime().Subtract(epoch).TotalSeconds;
+            return (long)time.ToUniversalTime().Subtract(epoch).TotalSeconds;
         }
     }
 }
