+++ HomegearLib.NET/Interface.cs
@@ -19,8 +19,8 @@ namespace HomegearLib
         private bool _default = false;
         public bool Default { get { return _default; } internal set { _default = value; } }
 
-        private int _physicalAddress = 0;
-        public int PhysicalAddress { get { return _physicalAddress; } internal set { _physicalAddress = value; } }
+        private long _physicalAddress = 0;
+        public long PhysicalAddress { get { return _physicalAddress; } internal set { _physicalAddress = value; } }
 
         private string _ipAddress = "";
         public string IpAddress { get { return _ipAddress; } internal set { _ipAddress = value; } }
@@ -28,11 +28,11 @@ namespace HomegearLib
         private string _hostname = "";
         public string Hostname { get { return _hostname; } internal set { _hostname = value; } }
 
-        private int _lastPacketReceived = 0;
-        public int LastPacketReceived { get { return _lastPacketReceived; } internal set { _lastPacketReceived = value; } }
+        private long _lastPacketReceived = 0;
+        public long LastPacketReceived { get { return _lastPacketReceived; } internal set { _lastPacketReceived = value; } }
 
-        private int _lastPacketSent = 0;
-        public int LastPacketSent { get { return _lastPacketSent; } internal set { _lastPacketSent = value; } }
+        private long _lastPacketSent = 0;
+        public long LastPacketSent { get { return _lastPacketSent; } internal set { _lastPacketSent = value; } }
 
         public Interface(Family family, string id, string type)
         {
